f8 roofline: make utils reusable
Summary:

Moves the float8 roofline gemm and memory traffic utils to
`torchao.float8.roofline_utils`, so they can be reused in other
places.

For now, I want to use this in ao_benchmarks.

Test Plan:

```
python benchmarks/float8/float8_roofline.py ~/local/tmp/test.txt --gemm_time_strategy roofline
```

Reviewers:

Subscribers:

Tasks:

Tags:
